+++ pkg/core/gasp/core/storage.go
@@ -8,7 +8,7 @@ import (
 )
 
 type GASPStorage interface {
-	FindKnownUTXOs(ctx context.Context, since uint64) ([]*overlay.Outpoint, error)
+	FindKnownUTXOs(ctx context.Context, since uint32) ([]*overlay.Outpoint, error)
 	HydrateGASPNode(ctx context.Context, graphID *overlay.Outpoint, outpoint *overlay.Outpoint, metadata bool) (*GASPNode, error)
 	FindNeededInputs(ctx context.Context, tx *GASPNode) (*GASPNodeResponse, error)
 	AppendToGraph(ctx context.Context, tx *GASPNode, spentBy *chainhash.Hash) error
